Decrease Stop_Distance As Lead Pulls Away
This rapidly decreases the following distance when the lead pulls away,
resulting in an early demand for acceleration.  This is helpful because the
MPC often stops ~4m behind the lead.  Meaning the lead has to move ~2m before the
MPC even considers requesting acceleration.
The offset is capped at half the total Stop_Distance, is linearly proportional to
the speed differential between ego and lead, requires the lead to be traveling
faster than Ego, and tapers away to 0 at 10m/s, all for safety and comfort.